+++ tests/data/full.atom.py
@@ -4,6 +4,7 @@ from reader import Content
 from reader import Enclosure
 from reader import EntrySource
 from reader._types import EntryData
+from reader.types import Author
 from reader._types import FeedData
 
 feed = FeedData(
@@ -11,7 +12,7 @@ feed = FeedData(
     updated=datetime.datetime(2003, 12, 13, 18, 30, 2, tzinfo=datetime.UTC),
     title='Example Feed',
     link='http://example.org/',
-    author='John Doe',
+    authors=(Author(name='John Doe'),),
     subtitle='all your examples are belong to us',
     version='atom10',
 )
@@ -23,7 +24,7 @@ entries = [
         updated=datetime.datetime(2003, 12, 13, 18, 30, 2, tzinfo=datetime.UTC),
         title='Atom-Powered Robots Run Amok',
         link='http://example.org/2003/12/13/atom03',
-        author='John Doe',
+        authors=(Author(name='John Doe'),),
         published=datetime.datetime(2003, 12, 13, 17, 17, 51, tzinfo=datetime.UTC),
         summary='Some text.',
         content=(
@@ -60,7 +61,7 @@ entries = [
             updated=datetime.datetime(2003, 12, 13, 18, 30, 2, tzinfo=datetime.UTC),
             title='Source Title',
             link='http://example.org/source',
-            author='Source Author',
+            authors=(Author(name='Source Author'),),
             subtitle='source subtitle',
         ),
     ),